requirements.txt
# Core ML dependencies
opencv-python>=4.8.0
numpy>=1.24.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Web framework
Flask>=3.0.0
flask-cors>=4.0.0

# Production server (optional, for deployment)
gunicorn>=21.2.0

